requirements.txt
Scrapy==0.24.6
Twisted==15.1.0
cffi==0.9.2
characteristic==14.3.0
cryptography==0.8.2
cssselect==0.9.1
enum34==1.0.4
ghp-import==0.4.1
lxml==3.4.4
mackup==0.7.4
pathlib==1.0.1
psutil==2.2.1
pyOpenSSL==0.15.1
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.12
pymongo==3.0.1
queuelib==1.2.2
redis==2.10.3
scrapy-mongodb==0.9.0
service-identity==14.0.0
six==1.9.0
thefuck==1.12
w3lib==1.11.0
wsgiref==0.1.2
zope.interface==4.1.2
pyexecjs
